move resources out of cogs
