+++ mushmom/resources/discord.yaml
